Learn entries with the prefix in them (#43)
This fixes a bug where rawrbot was unable to store entries that had the
bot plugin prefix in them. I was loading the prefix into the
configuration incorrectly so it only checked if the prefix was in the
message, when it should have checked if it was at the beginning of the
line.

Also, tests were configuring the prefix incorrectly, so the mock bot
wasn't respecting that configuration option. That's fixed as well.