Fix rubocop warnings and errors
I eliminated the rubocop_todo.yml file which had a bunch of stuff I
didn't want to deal with at the time. I also either fixed the
corresponding warnings in the code or deleted that code if I was
pretty sure it was either unused or broken. Added a few things to the
main rubocop.yml config file that seemed too onerous to fix or
described style things that I don't care about.