Fix Rubocop cop violations
This is a massive refactor to get everything in compliance with what
Rubocop declares is good Ruby style. This should give me a good baseline
for maintaining this style from here on out rather than just ignoring a
bunch of cops.

I explicitly ignored all the metrics cops that showed violations,
because I think they're too draconian as a general rule. I prefer to
judge complexity on a case-by-case basis, without using an arbitrary
algorithm and numeric value for it.