Limit reply length on learning plugin
People are letting entries get really really long, which results in them
spanning several irc messages, and causing the bot to run up against the
server's rate limiting and reply slowly. This adds a max message size
limit so we encourage folks to archive entries when they get too long.

The 'literal' command will still return the whole entry, so that users
have a way to easily get the full text of an entry.